+++ app/assets/stylesheets/react/form.scss
@@ -1,3 +1,3 @@
-.item-form {
-  padding-top: 93px;
+.modal-body {
+
 }
