+++ app/assets/stylesheets/react/index.scss
@@ -1,6 +1,8 @@
 @import url('https://fonts.googleapis.com/css?family=Permanent+Marker&display=swap');
 
 .navbar {
+  position: fixed;
+  width: 100%;
   z-index: 2;
   background: rgba(255, 255, 255, 0.5);
   transition: all 0.3s ease;
@@ -13,7 +15,6 @@
 }
 
 #app {
-  position: fixed;
   top: 0px;
   left: 0px;
   width: 100%;
@@ -41,6 +42,7 @@
     }
   }
 }
+
 
 .card-title {
   text-align: center;
@@ -78,6 +80,9 @@
 
 .mr-auto {
   margin: auto;
+}
+
+.index-btn {
 }
 
 // .navbar {
