+++ app/assets/stylesheets/react/show.scss
@@ -2,6 +2,10 @@
   margin-top: 5px;
 }
 
+.item-show-wrap {
+  position: relative;
+}
+
 .card-img {
   width: 100%;
   height: 25vh;
