+++ app/views/layouts/application.html.erb
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <title>swapx</title>
+  <title>SwapX</title>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <%= stylesheet_link_tag 'application', media: 'all' %>
   <link href="https://unpkg.com/nes.css@latest/css/nes.min.css" rel="stylesheet" />
@@ -11,25 +11,25 @@
   <nav class="navbar navbar-expand-lg">
     <div class="container-fluid">
       <h2 class="logo-text"><a href="/">SwapX</a></h2>
-      <div class="row">
-        <ul class="navbar-nav mr-auto">
-          <li class="nav-item nav-link">
-            <%= link_to 'My Profile', "/" %>
-          </li>
-          <li class="nav-item nav-link">
-            <%= link_to 'New Item', items_new_path %>
-          </li>
-        </ul>
-      </div>
       <div>
         <ul class="navbar-nav text-right">
+          <li class="nav-item mr-auto nav-link">
+            <%= link_to 'My Profile', "/" %>
+          </li>
+          <li class="nav-item mr-auto nav-link">
+            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#itemModal">
+              Add Item
+            </button>
+          </li>
           <%- if current_user -%>
           <li class="nav-item mr-auto">
             <%= link_to 'Sign Out', destroy_user_session_path, method: :delete, :class => "nav-link" %>
           </li>
           <%- else -%>
           <li class="nav-item btn btn-primary btn-sm sign-btns">
-            <%= link_to 'Sign Up', new_user_registration_path, :class => "nav-link" %>
+            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModalCenter">
+              Sign Up
+            </button>
           </li>
           <li class="nav-item btn btn-primary btn-sm sign-btns">
             <%= link_to 'Sign In', new_user_session_path, :class => "nav-link" %>
